+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/domain/nightstudy/repository/nightStudy/NightStudyRepository.kt
@@ -1,4 +1,4 @@
-package com.b1nd.dodamdodam.nightstudy.domain.nightstudy.repository
+package com.b1nd.dodamdodam.nightstudy.domain.nightstudy.repository.nightStudy
 
 import com.b1nd.dodamdodam.nightstudy.domain.nightstudy.entity.NightStudyEntity
 import org.springframework.data.jpa.repository.JpaRepository
