+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/domain/nightstudy/repository/nightStudyMember/NightStudyMemberQueryRepository.kt
@@ -1,4 +1,4 @@
-package com.b1nd.dodamdodam.nightstudy.domain.nightstudy.repository
+package com.b1nd.dodamdodam.nightstudy.domain.nightstudy.repository.nightStudyMember
 
 import com.b1nd.dodamdodam.nightstudy.domain.nightstudy.entity.NightStudyEntity
 import java.util.UUID
